requirements.txt
torch==2.3.0
torchvision==0.18.0
pillow==11.1.0
easydict==1.9
torchmetrics==1.4.0.post0
numpy==1.26.4
PyYAML==6.0
einops==0.6.0
tqdm==4.66.5
timm==0.9.7